+++ src/main/java/org/gridsuite/modification/server/entities/equipment/modification/LineSplitWithVoltageLevelEntity.java
@@ -113,10 +113,11 @@ public class LineSplitWithVoltageLevelEntity  extends ModificationEntity {
         );
     }
 
+    @Override
     public void setIdsToNull() {
         if (this.getMayVoltageLevelCreation() != null) {
             this.getMayVoltageLevelCreation().setId(null);
         }
-        this.setId(null);
+        super.setIdsToNull();
     }
 }
